database integrated with creating services by node manager

+++ Node-Manager/internal-api/routers/deployement.py
@@ -7,27 +7,34 @@ import requests
 import uvicorn
 from decouple import config
 from fastapi import APIRouter, BackgroundTasks, Depends, File, HTTPException, UploadFile
+from pymongo import MongoClient
 
 from utils.jwt_bearer import JWTBearer
 from utils.storage import uploadFile
 from utils.verify_zip import verify_zip
 
-sys.path.append("../")
+sys.path.append("..")
 
 router = APIRouter()
 
 
 CONTAINER_NAME = config("deploy_app_container_name")
+mongokey = config("mongoKey")
+client = MongoClient(mongokey)
+db = client["services"]
+collection = db.services
 
 
-async def my_task(time: int, file: UploadFile = File(...)):
+async def schedule_deployement_task(time: int, file: UploadFile = File(...)):
     await asyncio.sleep(time)
     # Logic or api call will come here to deploy
     try:
-        print("hello")
+        node = collection.find_one({"name": "node-manager"})
+        if not node:
+            return
         fname = file.filename
         fname = fname.split(".")[0]
-        res = requests.post(f"http://127.0.0.1:8001/deploy/{fname}")
+        res = requests.post(f"http://{node['ip']}:{node['port']}/deploy/{fname}")
         os.remove(file.filename)
         print(status)
         print(res.text)
@@ -54,9 +61,12 @@ async def upload_zip_file(file: UploadFile = File(...)):
     if verify_zip(f"{file.filename}"):
         # Upload to the cloud
         status = uploadFile(CONTAINER_NAME, ".", file.filename)
+        node = collection.find_one({"name": "node-manager"})
+        if not node:
+            return
         fname = file.filename
         fname = fname.split(".")[0]
-        res = requests.post(f"http://127.0.0.1:8001/deploy/{fname}")
+        res = requests.post(f"http://{node['ip']}:{node['port']}/deploy/{fname}")
         os.remove(file.filename)
         print(status)
         print(res.text)
@@ -82,7 +92,7 @@ async def schedule_task(background_tasks: BackgroundTasks, time: int = 0, file: 
         status = uploadFile(CONTAINER_NAME, ".", file.filename)
         fname = file.filename
         fname = fname.split(".")[0]
-        background_tasks.add_task(my_task, time, file)
+        background_tasks.add_task(schedule_deployement_task, time, file)
         return {"message": "Task scheduled", "status": json.dumps(status)}
     else:
         os.remove(file.filename)
